test/requirements.txt
pyserial==3.5
pytest==7.4.0
paho-mqtt=1.6.1
